Logo_Detection_Deployment/requirements.txt
numpy==1.19.5
tensorflow==2.5.0
fastapi==0.65.1
Pillow==8.2.0
typing-extensions==4.2.0
starlette==0.14.2
torch==1.11.0
pillow==8.2.0